修JEI兼容崩溃

+++ ConfluenceOtherworld/src/main/java/org/confluence/mod/common/event/game/entity/LivingEntityEvents.java
@@ -229,7 +229,7 @@ public final class LivingEntityEvents {
         // 暴击判定和伤害显示
         boolean crit = false;
         if (!TCAttributes.hasCustomAttribute(TCAttributes.CRIT_CHANCE) && attacker instanceof Player player) {
-            if (LibUtils.checkChance((float) player.getAttributeValue(TCAttributes.CRIT_CHANCE), player.getRandom())) {
+            if (LibUtils.checkChance(player.getAttributeValue(TCAttributes.CRIT_CHANCE), player.getRandom())) {
                 amount *= 1.5F;
                 player.crit(victim);
                 crit = true;
